Deduplicate browser auth functions
